getting webauthn to work, WIP
